requirements.txt

### **d. `requirements.txt`**

Lists all Python dependencies required for the project.

```plaintext
# requirements.txt

pandas==1.5.3
numpy==1.24.2
nltk==3.8.1
vaderSentiment==3.3.2
textblob==0.17.1
gensim==4.3.1
spacy==3.5.1
matplotlib==3.7.1
seaborn==0.12.2
wordcloud==1.8.2
scikit-learn==1.2.2
python-dotenv==1.0.0
pyLDAvis==3.4.0
emoji==2.4.0
plotly==5.15.0
dash==2.15.0
streamlit==1.26.0
transformers==4.30.0
rake-nltk==1.0.7
yake==0.4.4
swifter==1.0.6
papermill==2.4.0
pytest==7.2.0
google-api-python-client==2.98.0
